[WIP] Create TodoList module
Here we create the interface for manipulating the todo lists and we also
add documentation and typespecs to it, as well as tests.

This is ultra importante because people need to understand how this code
works without looking at its source. Or even with the source code beside
them, they should be able to understand what's going on without looking
at the inner workings.